+++ src/angular-ladda.js
@@ -24,18 +24,26 @@
     .directive('ladda', ['$compile', '$timeout', 'ladda', function ($compile, $timeout, laddaOption) {
       return {
         restrict: 'A',
-        replace: false,
-        terminal: true,
-        priority: 501,
+        priority: -1,
         link: function (scope, element, attrs) {
           element.addClass('ladda-button');
           if(angular.isUndefined(element.attr('data-style'))) {
             element.attr('data-style', laddaOption.style || 'zoom-in');
           }
+
+          // ladda breaks childNode's event property.
+          // because ladda use innerHTML instead of append node
+          if(!element[0].querySelector('.ladda-label')) {
+            var labelWrapper = document.createElement('span');
+            labelWrapper.className = 'ladda-label';
+            angular.element(labelWrapper).append(element.contents());
+            element.append(labelWrapper);
+          }
+
+          // create ladda button
           var ladda = Ladda.create( element[0] );
 
-          $compile(element, null, 501)(scope);
-
+          // add watch!
           scope.$watch(attrs.ladda, function(loading) {
             if(loading || angular.isNumber(loading)) {
               if(!ladda.isLoading()) {